Cleanup before reporting refactoring
